"Chore: Separate the css into another file"
